+++ package.json
@@ -9,16 +9,20 @@
     "lint:fix": "eslint --fix **/*.js"
   },
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "cors": "2.8.5",
     "cross-env": "7.0.3",
     "dotenv": "^9.0.1",
     "express": "4.17.1",
     "express-validator": "^6.10.1",
     "joi": "^17.4.0",
+    "jsonwebtoken": "8.5.1",
     "lowdb": "^1.0.0",
     "mongodb": "^3.6.6",
     "mongoose": "^5.12.7",
     "morgan": "1.10.0",
+    "passport": "0.4.1",
+    "passport-jwt": "4.0.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
